dev: Run jekyll through container
This patch adds a Dockerfile to create a container to run jekyll and a
`jekyll.sh` which builds and start the image, isolating the build from
the development system.

The start script will create a directory '.vendor/bundle' in the project
directory that will be used to hold the necessary ruby gems and support
files.

The first run will take a lot longer, as the image wil be built, the
bundle will be installed (and all gems will be downloaded and compiled),
before 'jekyll serve' is called.

Note that jekyll will serve for any address ('--host 0.0.0.0') on port
4000, and this port should be firewalled on the host machine.